+++ src/presenters/pages/search.js
@@ -7,13 +7,31 @@ import NotFound from 'Components/errors/not-found';
 import Layout from '../layout';
 import MoreIdeas from '../more-ideas';
 
-const SearchPage = ({ query }) => (
-  <Layout searchQuery={query}>
-    {!!query && <Helmet title={`Search for ${query}`} />}
-    {query ? <SearchResults query={query} /> : <NotFound name="anything" />}
-    <MoreIdeas />
-  </Layout>
-);
+import { useAlgoliaSearch, useLegacySearch } from '../../state/search';
+import useDevToggle from '../includes/dev-toggles';
+
+// Hooks can't be _used_ conditionally, but components can be _rendered_ conditionally
+const AlgoliaSearchWrapper = ({ query }) => {
+  const searchResults = useAlgoliaSearch(query);
+  return <SearchResults query={query} searchResults={searchResults} />;
+};
+
+const LegacySearchWrapper = ({ query }) => {
+  const searchResults = useLegacySearch(query);
+  return <SearchResults query={query} searchResults={searchResults} />;
+};
+
+const SearchPage = ({ query }) => {
+  const algoliaFlag = useDevToggle('Algolia Search');
+  const SearchWrapper = algoliaFlag ? AlgoliaSearchWrapper : LegacySearchWrapper;
+  return (
+    <Layout searchQuery={query}>
+      {!!query && <Helmet title={`Search for ${query}`} />}
+      {query ? <SearchWrapper query={query} /> : <NotFound name="anything" />}
+      <MoreIdeas />
+    </Layout>
+  );
+};
 
 SearchPage.propTypes = {
   query: PropTypes.string,
